backend_django_dev/requirements.txt
Django>=5.2,<6.0
python-dotenv>=1.0,<2.0
uvicorn>=0.35,<1.0
websockets>=15.0,<16.0
psycopg2-binary>=2.9,<3.0
celery>=5.6,<6.0
django-celery-beat>=2.8,<3.0
redis>=5.2,<6.0
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.4,<6.0
django-cors-headers>=4.8,<5.0
channels>=4.2,<5.0
channels-redis>=4.3,<5.0
django-extensions>=4.1,<5.0
pydot>=4.0,<5.0
dj-database-url>=2.3,<3.0
drf-spectacular>=0.28,<1.0
requests>=2.32,<3.0
requests-oauthlib>=2.0,<3.0
openai>=1.0,<2.0
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
